Feat: Adapt Terragrunt code for aws and k3d
This is a monster commit with many changes that were required for the
repo to be able to serve both aws and k3d with max code sharing. There
are many aspects to what's going on but some major points would be:

1- Update devcontainer for the required features
2- Divide tf repos to logical chunks that can be applied to multiple
targets. As an example, services tf repo deals with creating services
and it works exactly the same way for both aws and k3d.
3- Add k3d testing for github actions.
4- Create a separation between modules and targets. Targets are the
leaves that interact with the environment at which deployment is
happening while modules capture all the common behavior any target would
accept. Every module has at least one target that inherits its
configuration.